[FEATURE] Refactor the score progression into the stats app.
